+++ utilities/carp_study_generator/CHANGELOG.md
@@ -1,8 +1,9 @@
-## 0.30.3
+## 0.30.3+2
 * upgrade to null-safety
 * more robust error messages in `dryrun`
 * upgrade to `carp_backend` v 0.30.3
 * support for setting study descriptions
+* fix of small errors
 
 ## 0.21.5
 * improvement to `localization` command